+++ Simulation.h
@@ -33,9 +33,9 @@ public:
 	void core();
 	void saveFrame(fstream& outFile, Particle* particles, int particleNumber);
 	void saveProgress(char* ofName, Particle* particles, int particleNumber, int frameStep);
-	double random(double min, double max);
+	float random(float min, float max);
 	void printProgress(int currentFrame, int totalFrames);
-	void integrator(Particle* particles, int NPARTICLE, int FRAMESTEP, int N, double* coefc, double* coefd);
+	void integrator(Particle* particles, int NPARTICLE, int FRAMESTEP, int N, float* coefc, float* coefd);
 	Particle* getParticle() const;
 	int getNParticle() const;
 	int* getFrameRef();
